+++ Realization/backend/core/core-api/src/main/java/eu/bcvsolutions/idm/core/api/dto/filter/IdmAutomaticRoleAttributeRuleFilter.java
@@ -86,6 +86,4 @@ public class IdmAutomaticRoleAttributeRuleFilter extends DataFilter {
 	public void setComparison(AutomaticRoleAttributeRuleComparison comparison) {
 		this.comparison = comparison;
 	}
-	
-	
 }