+++ Realization/backend/core/core-impl/src/main/java/eu/bcvsolutions/idm/core/model/service/impl/DefaultIdmIdentityRoleService.java
@@ -229,6 +229,13 @@ public class DefaultIdmIdentityRoleService
 					);
 		}
 		//
+		if (filter.getIdentityContractId() != null) {
+			predicates.add(builder.equal(
+					root.get(IdmIdentityRole_.identityContract).get(AbstractEntity_.id), 
+					filter.getIdentityContractId())
+					);
+		}
+		//
 		return predicates;
 	}