+++ Realization/backend/core/core-impl/src/main/java/eu/bcvsolutions/idm/core/scheduler/task/impl/ProcessAutomaticRoleByAttributeTaskExecutor.java
@@ -13,6 +13,7 @@ import org.springframework.stereotype.Service;
 
 import com.google.common.collect.Sets;
 
+import eu.bcvsolutions.idm.core.api.domain.ContractState;
 import eu.bcvsolutions.idm.core.api.domain.CoreResultCode;
 import eu.bcvsolutions.idm.core.api.dto.AbstractIdmAutomaticRoleDto;
 import eu.bcvsolutions.idm.core.api.dto.IdmAutomaticRoleAttributeDto;
@@ -27,7 +28,7 @@ import eu.bcvsolutions.idm.core.api.service.IdmIdentityRoleService;
 import eu.bcvsolutions.idm.core.api.service.IdmRoleRequestService;
 
 /**
- * Process all identities that passed and not passed given automatic role
+ * Process all contracts that passed and not passed given automatic role
  * 
  * @author Ondrej Kopr <kopr@xyxy.cz>
  *
@@ -59,52 +60,62 @@ public class ProcessAutomaticRoleByAttributeTaskExecutor extends AbstractAutomat
 		}
 		Set<AbstractIdmAutomaticRoleDto> setWithAutomaticRole = Sets.newHashSet(automaticRolAttributeDto);
 		//
-		Page<UUID> newPassedIdentities = automaticRoleAttributeService.getNewPassedIdentitiesForAutomaticRole(automaticRoleId, new PageRequest(0, PAGE_SIZE));
-		//
-		Page<UUID> newNotPassedIdentities = automaticRoleAttributeService.getNewNotPassedIdentitiesForAutomaticRole(automaticRoleId, new PageRequest(0, PAGE_SIZE));
+		// by contract
+		Page<UUID> newPassedContracts = automaticRoleAttributeService.getContractsForAutomaticRole(automaticRoleId, true, new PageRequest(0, PAGE_SIZE));
+    	Page<UUID> newNotPassedContracts = automaticRoleAttributeService.getContractsForAutomaticRole(automaticRoleId, false, new PageRequest(0, PAGE_SIZE));
 		//
 		counter = 0L;
-		count = Long.valueOf(newPassedIdentities.getTotalElements() + newNotPassedIdentities.getTotalElements());
+		count = Long.valueOf(newNotPassedContracts.getTotalElements() + newNotPassedContracts.getTotalElements());
 		//
 		// assign new passed roles
 		boolean canContinue = true;
+    	//
+    	// process contracts
+    	canContinue = true;
     	while (canContinue) {
-    		for(UUID identityId : newPassedIdentities) {
-    			IdmIdentityContractDto primeContract = identityContractService.getPrimeContract(identityId);
-    			IdmRoleRequestDto roleRequest = automaticRoleAttributeService.prepareAddAutomaticRoles(primeContract, setWithAutomaticRole);
-    			roleRequestService.startRequestInternal(roleRequest.getId(), false);
-    			counter++;
-    			canContinue = updateState();
-    			if (!canContinue) {
-    				break;
+    		for(UUID contractId : newPassedContracts) {
+    			IdmIdentityContractDto contract = identityContractService.get(contractId);
+    			//
+    			// check for contract validity
+    			if (contract.getState() == ContractState.DISABLED || !contract.isValidNowOrInFuture()) {
+    				continue;
     			}
+    			//
+				IdmRoleRequestDto roleRequest = automaticRoleAttributeService.prepareAddAutomaticRoles(contract, setWithAutomaticRole);
+				roleRequestService.startRequestInternal(roleRequest.getId(), false);
+				//
+				counter++;
+				canContinue = updateState();
+				if (!canContinue) {
+					break;
+				}
     		}
-    		if (newPassedIdentities.hasNext()) {
-    			newPassedIdentities = automaticRoleAttributeService.getNewPassedIdentitiesForAutomaticRole(automaticRoleId, newPassedIdentities.nextPageable());
+    		if (newPassedContracts.hasNext()) {
+    			newPassedContracts = automaticRoleAttributeService.getContractsForAutomaticRole(automaticRoleId, true, newPassedContracts.nextPageable());
     		} else {
     			break;
     		}
     	}
     	//
-    	// remove new not passed roles
     	while (canContinue) {
-    		for(UUID identityId : newNotPassedIdentities) {
+    		for(UUID contractId : newNotPassedContracts) {
     			IdmIdentityRoleFilter filter = new IdmIdentityRoleFilter();
-    			filter.setIdentityId(identityId);
+    			filter.setIdentityContractId(contractId);
     			filter.setAutomaticRoleId(automaticRoleId);
     			List<IdmIdentityRoleDto> identityRoles = identityRoleService.find(filter, null).getContent();
     			for (IdmIdentityRoleDto identityRole : identityRoles) {
     				IdmRoleRequestDto roleRequest = automaticRoleAttributeService.prepareRemoveAutomaticRoles(identityRole, setWithAutomaticRole);
     				roleRequestService.startRequestInternal(roleRequest.getId(), false);
     			}
+    			//
     			counter++;
     			canContinue = updateState();
     			if (!canContinue) {
     				break;
     			}
     		}
-    		if (newNotPassedIdentities.hasNext()) {
-    			newNotPassedIdentities = automaticRoleAttributeService.getNewNotPassedIdentitiesForAutomaticRole(automaticRoleId, newNotPassedIdentities.nextPageable());
+    		if (newNotPassedContracts.hasNext()) {
+    			newNotPassedContracts = automaticRoleAttributeService.getContractsForAutomaticRole(automaticRoleId, false, newNotPassedContracts.nextPageable());
     		} else {
     			break;
     		}