+++ Realization/backend/core/core-impl/src/main/resources/eu/bcvsolutions/idm/core/sql/postgresql/V8_00_005__fix-enum-name-automatic-role.sql
@@ -1,0 +1,7 @@
+--
+-- CzechIdM 8 Flyway script 
+-- BCV solutions s.r.o.
+--
+-- fix enum name for AutomaticRoleAttributeRuleType
+
+UPDATE idm_auto_role_att_rule SET type = 'IDENTITY_EAV' where type = 'IDENITITY_EAV';