+++ Realization/backend/core/core-impl/src/test/java/eu/bcvsolutions/idm/core/scheduler/TestSchedulableDryRunTask.java
@@ -1,0 +1,15 @@
+package eu.bcvsolutions.idm.core.scheduler;
+
+/**
+ * Test task for {@link DefaultSchedulerManagerIntegrationTest}.
+ * 
+ * @author Radek Tomiška
+ *
+ */
+public class TestSchedulableDryRunTask extends TestSchedulableTask {
+	
+	@Override
+	public boolean supportsDryRun() {
+		return true;
+	}
+}