+++ Realization/frontend/czechidm-core/src/content/workflow/Definitions.js
@@ -1,5 +1,4 @@
 import React from 'react';
-import Helmet from 'react-helmet';
 import { connect } from 'react-redux';
 import * as Basic from '../../components/basic';
 import * as Services from '../../services';